port capacity int to double

+++ Data/Entities/IpAssignment.cs
@@ -4,7 +4,7 @@ namespace Data.Entities;
 public class IpAssignment : BaseEntity
 {
     public IpAssignmentTypes Type { get; set; }
-    public int? Capacity { get; set; }
+    public double? Capacity { get; set; }
 
     public int IpAddressId { get; set; }
     public IpAddress? IpAddress { get; set; }
